rag/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
blinker==1.7.0
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
coloredlogs==15.0.1
cryptography==42.0.5
dataclasses-json==0.6.4
deprecated==1.2.14
fastapi==0.110.2
fastembed==0.2.6
filelock==3.13.4
flask==3.0.3
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.3.1
google-auth==2.29.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.2
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.7
importlib-metadata==7.0.0
importlib-resources==6.4.0
itsdangerous==2.2.0
jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
kubernetes==29.0.0
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.45
langchain-text-splitters==0.0.1
langsmith==0.1.49
loguru==0.7.2
markdown-it-py==3.0.0
markupsafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
onnx==1.16.0
onnxruntime==1.17.3
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.1
overrides==7.7.0
packaging==23.2
pdfminer-six==20231228
pdfplumber==0.11.0
pillow==10.3.0
posthog==3.5.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pulsar-client==3.5.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
pydantic==2.7.0
pydantic-core==2.18.1
pygments==2.17.2
pypdfium2==4.29.0
pypika==0.48.9
pyproject-hooks==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyyaml==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlalchemy==2.0.29
starlette==0.37.2
sympy==1.12
tenacity==8.2.3
tokenizers==0.15.2
tqdm==4.66.2
typer==0.12.3
typing-extensions==4.11.0
typing-inspect==0.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
werkzeug==3.0.2
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.1